leverage sideeffects of Overnightable for CopyWeek
CopyWeek ignores the next days of overnightables. It dups the first
half, but Scheduling#mode_to_week_and_year sets the end date from the
next day, so the copied next day is created through magic hooks in
Overnightable.

Moving a persisted Scheduling on the other hand must first remove its
next day, because the end_* creates it magically for us.

As noted, on the long run, we have to overthink the architecture of the
overnightables. They bleed in every feature. Every buisiness ibject has
to deal with them in a special way.

This has to stop.