overnightable splits at view layer
For this, we use a #focus_day, which tells the decorated Schedulable on
which day it is currently rendered (nightshifts have two).

Shifts have a day, for example number 3;
Schedulings have dates that must match the day.

Some copying involved, but this layer issupposed to go woosh when
switching to Ember.

*FilterDecorator#schedulings_for returns decorated records.

Note: wwt may count overnightables twice because it is present in two
weeks, basically. Asking...