shift down schedulings when una is in cell
This avoids ugly overlapping and replaces it with just a slight overlap,
so the plan user can read the time/all_day and the reason (description
is not shown anyway) of Unas even if they are Schedulings in the same
cell.

Caveat: a second Una would be covered totally and look ugly again. I see
no way to fix this currently.